+++ crates/bitwarden-user-crypto-management/src/user_crypto_management_client.rs
